Lexer Parser + namechange

+++ DropletsInMotion/Communication/Simulator/Models/ActionQueueItem.cs
@@ -4,7 +4,7 @@ namespace DropletsInMotion.Communication.Simulator.Models
 {
     public class ActionQueueItem
     {
-        public ActionQueueItem(ActionItem action, decimal time)
+        public ActionQueueItem(ActionItem action, double time)
         {
             Action = action;
             Time = time;
@@ -14,7 +14,7 @@ namespace DropletsInMotion.Communication.Simulator.Models
         public ActionItem Action { get; set; }
 
         [JsonPropertyName("time")]
-        public decimal Time { get; set; }
+        public double Time { get; set; }
 
         public override string ToString()
         {
